[CL-1941] Move all widgets into shared folder
