+++ front/app/components/ContentBuilder/Widgets/AboutBox/index.tsx
