+++ front/app/components/ContentBuilder/Widgets/AboutBox/messages.ts
