+++ front/app/components/ContentBuilder/Widgets/Accordion/index.tsx
