+++ front/app/components/ContentBuilder/Widgets/Accordion/messages.ts
