+++ front/app/components/ContentBuilder/Widgets/Image/messages.ts
