+++ front/app/components/ContentBuilder/Widgets/Text/index.tsx
