+++ front/app/components/ContentBuilder/Widgets/TwoColumn/index.tsx
