+++ front/app/components/ContentBuilder/Widgets/TwoColumn/messages.ts
