+++ front/app/components/ContentBuilder/Widgets/WhiteSpace/index.tsx
