+++ front/app/components/ContentBuilder/Widgets/WhiteSpace/messages.ts
