+++ front/app/components/ContentBuilder/constants.ts
@@ -1,0 +1,4 @@
+export const IMAGE_UPLOADING_EVENT = 'imageUploading';
+export const CONTENT_BUILDER_ERROR_EVENT = 'contentBuilderError';
+export const CONTENT_BUILDER_DELETE_ELEMENT_EVENT =
+  'deleteContentBuilderElement';
