+++ front/app/components/ContentBuilder/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  urlPlaceholder: {
+    id: 'app.containers.admin.ContentBuilder.urlPlaceholder',
+    defaultMessage: 'https://example.com',
+  },
+});
